Renamed src/js/dashboard-v3 to src/js/dashboard
