+++ pages/_includes/footer.njk
@@ -147,7 +147,7 @@ ga('tracker3.send', 'pageview');
 
 {%- pagesectionincludes page, "state-dashboard" %}
 	<script type="module" src="https://d3js.org/d3.v6.min.js"></script><!-- had some trouble importing this, error circular dependencies after npm install so including here -->
-	{{ macros.includeJavascriptByEnv('dashboard-v3', env.dev) }}
+	{{ macros.includeJavascriptByEnv('dashboard', env.dev) }}
 {% endpagesectionincludes -%}
 
 {%- pagesection page, "equity" %}
