+++ src/js/charts-sandbox/index.js
@@ -4,16 +4,16 @@ console.log("Loading chart sandbox");
 import '../common/search/index.js';
 import '../common/county-buttons/buttons.js'
 
-import "../dashboard-v3/charts/cagov-chart-dashboard-confirmed-cases/index.js"
+import "../dashboard/charts/cagov-chart-dashboard-confirmed-cases/index.js"
 
 // test version of deaths chart using HighCharts
 import "./charts/cagov-chart-dashboard-confirmed-cases-hc/index.js"
 // import "./charts/cagov-chart-dashboard-confirmed-deaths/index.js"
 
-import "../dashboard-v3/charts/cagov-chart-dashboard-total-tests/index.js"
-import "../dashboard-v3/charts/cagov-chart-dashboard-positivity-rate/index.js"
+import "../dashboard/charts/cagov-chart-dashboard-total-tests/index.js"
+import "../dashboard/charts/cagov-chart-dashboard-positivity-rate/index.js"
 
-import "../dashboard-v3/charts/cagov-chart-dashboard-patients/index.js"
-import "../dashboard-v3/charts/cagov-chart-dashboard-icu-beds/index.js"
+import "../dashboard/charts/cagov-chart-dashboard-patients/index.js"
+import "../dashboard/charts/cagov-chart-dashboard-icu-beds/index.js"
 
 import '../common/search/filters.js'
