+++ src/js/dashboard/DOCS.md
