+++ src/js/dashboard/async-polyfill.js
