+++ src/js/dashboard/charts/cagov-chart-dashboard-confirmed-cases/index.js
