+++ src/js/dashboard/charts/cagov-chart-dashboard-confirmed-deaths/index.js
