+++ src/js/dashboard/charts/cagov-chart-dashboard-groups-age-deaths/index.scss
