+++ src/js/dashboard/charts/cagov-chart-dashboard-groups-gender-cases/index.scss
