+++ src/js/dashboard/charts/cagov-chart-dashboard-groups-gender-deaths/index.scss
