+++ src/js/dashboard/charts/cagov-chart-dashboard-icu-beds/index.js
