+++ src/js/dashboard/charts/cagov-chart-dashboard-patients/histogram.js
