+++ src/js/dashboard/charts/cagov-chart-dashboard-sparkline/sparkline-template.js
