+++ src/js/dashboard/charts/cagov-chart-dashboard-sparkline/sparkline.scss
