+++ src/js/dashboard/charts/common/histogram-template.js
