+++ src/js/dashboard/charts/common/histogram.js
