+++ src/js/dashboard/charts/common/histogram.scss
