+++ src/js/dashboard/charts/common/line-chart-config.json
