+++ src/js/dashboard/charts/ie11.scss
